hana-multidb-mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hana-multidb-mcp"
version = "0.1.0"
description = "Multi-connection HANA MCP server with GitHub-backed connection catalog"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "hdbcli>=2.19",
]

[project.scripts]
hana-multidb-mcp = "src.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
